SpeechToTextToSpeech/requirements.txt
Flask
langchain
chromadb
openai
pydub
speechrecognition
python-dotenv
